add ability to start task for SkyNet

+++ Trunk/TBox/Plugins/SkyNet/Code/TaskExecutor.cs
@@ -1,39 +1,52 @@
 ﻿using System.IO;
-using System.Windows;
 using Mnk.Library.ScriptEngine;
+using Mnk.Library.ScriptEngine.Core.Interfaces;
+using Mnk.TBox.Core.Contracts;
 using Mnk.TBox.Plugins.SkyNet.Code.Interfaces;
+using Mnk.TBox.Tools.SkyNet.Common;
+using ServiceStack.Text;
 
 namespace Mnk.TBox.Plugins.SkyNet.Code
 {
     class TaskExecutor : ITaskExecutor
     {
         private readonly IDataPacker dataPacker;
-        private readonly IServicesBuilder servicesBuilder;
+        private readonly IScriptCompiler<ISkyScript> compiler;
+        private readonly IPluginContext context;
+        private readonly IServicesFacade servicesFacade;
         private readonly IConfigsFacade configsFacade;
 
-        public TaskExecutor(IServicesBuilder servicesBuilder, IConfigsFacade configsFacade, IDataPacker dataPacker)
+        public TaskExecutor(IServicesFacade servicesFacade, IConfigsFacade configsFacade, IDataPacker dataPacker, IScriptCompiler<ISkyScript> compiler, IPluginContext context)
         {
-            this.servicesBuilder = servicesBuilder;
+            this.servicesFacade = servicesFacade;
             this.configsFacade = configsFacade;
             this.dataPacker = dataPacker;
+            this.compiler = compiler;
+            this.context = context;
         }
 
-        public void Execute(SingleFileOperation op)
+        public string Execute(SingleFileOperation operation)
         {
-            string name;
-            var path = dataPacker.Pack(
-                @"d:\sample.dll",
-                new[] { "*.dll" }, out name);
-            using (var cl = servicesBuilder.CreateFileServerClient(configsFacade.GetAgentConfig()))
+            var config = configsFacade.AgentConfig;
+            var scriptContent = File.ReadAllText(Path.Combine(context.DataProvider.ReadOnlyDataPath, operation.Path));
+            var script = compiler.Compile(scriptContent, operation.Parameters);
+            var task = new ServerTask
             {
-                using (var f = File.OpenRead(path))
-                {
-                    var id = cl.Instance.Upload(f);
-                    MessageBox.Show(id);
-                }
-            }
-            MessageBox.Show(path);
+                Owner = config.Name + ":" + config.Port,
+                Script = scriptContent,
+                ScriptParameters = JsonSerializer.SerializeToString(operation.Parameters),
+                ZipPackageId = PackData(script, config)
+            };
+            return servicesFacade.StartTask(config, task);
+        }
+
+        private string PackData(ISkyScript script, AgentConfig config)
+        {
+            if (string.IsNullOrEmpty(script.DataFolderPath)) return string.Empty;
+            var path = dataPacker.Pack(script.DataFolderPath, script.PathMasksToInclude);
+            var zipPackageId = servicesFacade.UploadFile(config, path);
             File.Delete(path);
+            return zipPackageId;
         }
     }
 }
